[sql] Update SQL types
After having time to consult the documentation, it's more flexible to
use datetime instead of timestamp (it avoids timezone issues) and ip
address and ttl can be represented in the more compact ascii character
set. This allows for display of the values in the CLI more easily as
well, versus raw varbinary. The binary encoded IP addresses should
still be stored as binary for the time being.