[main] Prefer absolute paths in golang
The golang dev team discourages relative imports, and they are not
well supported for portable builds. Return to absolute imports, which
work with go modules. The go build tools are smart enough to realize
this module provides com.github/uw-ictd/haulage and not actually go
out to github for all of them on build.