[main] Re-enable radius server
The radius server was being started in the foreground, which then
halted the main process before the worker processes and packet
processing begin. The radius server should run in its own
go-routine. This commit launches the entire radius server setup and
running in its own go-routine. This has the benefit of not halting the
main program if the radius setup fails.