Devel missing flowlogs (#21)
* [radius] Disable broken radius server

The radius server implementation was not actually well-tested and
breaks system operation when not operating in radius mode. Notably,
the radius server is initialized on the main run loop, and once it
starts will hang main operation if there is no radius traffic.

* [modules] Update dependencies

* [sql] Support unicode usernames

* [main] Add a log verbosity parameter to enable debug logging

* [classify][main] Parameterize user subnet configuration

Previously the address range attributed to users was hard-coded in the
source. This commit makes the subnet and any ignored addresses a part
of the configuration and changeable at runtime.