build: Specify phony targets
These targets are "commands" and not actual files which should be
built. They should be specified as phony to make knows to not check if
they have already been built when called. For example, without this
commit, making a directory "build" prevents the build command from
being called, since the build target will be considered up to date.