build: Support building and packaging multiple architectures
This commit re-works the makefile to allow building and packaging for
multiple architectures, regardless of what the host architecture
is. It does assume that if you wish to cross-compile for another
architecture, that you have the appropriate cross-compile toolchains
and libraries installed and ready. A dockerfile will be added shortly
that makes this easy to do in isolation.

This change goes through some length to ensure that basic native
compilation and packaging works transparently, allowing local usage
without a fancy cross-compilation setup.