pkg: Add support for building multi-arch multi-distro packages
This commit adds docker infrastructure to build packages for debian
and ubuntu using each distribution's multi-architecture toolchain. The
packages are build in isolated containers for each base distribution,
and then copied out into the main build directory. Docker is *not*
required to run the packages-- it's just used to build them instead of
with VMs.