aggregator: Initial async aggregator to collect usage in memory
The aggregator is implemented as a main dispatcher which processes
messages generated from each packet processing task. The dispatcher
owns a datastructure with the existing aggregation tasks, and forwards
reports to the appropriate aggregation task. This is heavyweight for
simple aggregation, but will be extended with per-aggregation timeouts
and database writes that will leverage the channel and task
infrastructure.