packet_parser: Expand returned data for more complete parsing
This commit embraces structuring the parse as a tree with each
function call into the next layer providing additional information
context from the previous layers, and the final return value created
at the leaf including all the parsed information thus far. Importantly
this approach does not return references to any of the inner
intermediate values, making it important to do each parsing step
within the tree. This might become cumbersome to extend in the future,
but is manageable now since the number of cases we are interested in is
limited.