main aggregator parser: Add initial flow normalization
In the haulage model, flows are either from a user to some remote
service, from a user to another user, or some kind of "other" flow
with no user involved. This commit adds logic to normalize arbitrary
flow fivetuples (which may have either a source or destination user)
to one of these more structured flow types. Right now this logic is
primarily implemented in the main module, but will be refactored out
in a forthcoming commit.