enforcer main: Add an initial policy enforcement engine
The enforcer currently is based on single user accounts. It handles
the task of tracking the decrease in the user's data balance, storing
the updated balance, and will eventually manage system forwarding
rules for each user.

The current implementation is simplistic and assumes users are
identified by a single IP address, which doesn't make sense in the
context of IPv4/IPv6 dual connectivity or link-local addresses.