main: Check database migration status before running
This commit adds logic to check if there are any outstanding or extra
migrations between the installed migrations and the current database
schema. If there is a difference, the program halts and does not
proceed. It also adds command options to specify the location of
the migrations files and to upgrade the database with available
migrations files.