migrations python: Remove currency concerns from haulage
Haulage is designed to only worry about the data balance, not the
monetary currency balance of each user. This commit removes the
monetary balance and currency from the haulage database to support
better divisions of concerns between haulage and CoLTE. This modifies
the initial migration without making a new migration file since the
migrations have not been released yet.