haulage: Re-work early migrations to avoid CoLTE conflict
As documented in https://github.com/uw-ictd/haulage/issues/25, both
CoLTE and haulage refer to the "currencies" table in the same postgres
database at different points in their schema history. Ownership of the
currencies table transitioned from haulage to CoLTE, and conflicts
would arise if some or all of the CoLTE migrations ran before the
haulage migrations removing their reference to the currencies table.

Unfortunatley since versions of both CoLTE and haulage have been
released with this conflict, it is not a simple task to just update
the haulage migrations to avoid the conflict. This commit includes
logic to "re-write the migration history" (DANGER!) to address this
issue in all 5 of the following cases:

1) Haulage ran first, then colte ran, database is healthy:

In this case the original init and remove currency balance migrations
will have been applied to the db, and the currencies table will exist,
owned by CoLTE. There will be "extra" migrations applied to the db
relative to the migrations present on disk in the haulage migrations
directory. The new "hack" migrations will then be applied, but have no
effect since they are all conditional on field/table existence. The
final "hack" migration will then remove the record of the original
migrations, to prevent future warnings about extra migrations (since
they have been effectively replaced by the hack migrations).

2) colte ran first, then haulage attempts to run, haulage migrate
fails:

In this case the current system state is unhealthy, and haulage is
unable to run since its migrations could not be applied. After this
commit though, the new "hack" migrations with existence conditions can
be applied cleanly. No history will be re-written by the final hack
migration since the earlier migrations failed to apply.

3) haulage has run, colte never ran: database is healthy

Similarly to case 1, the "hack" migrations will have no effect, and
the earlier broken non-hack migrations will be removed from history.

4) colte has run, haulage has never run: database is healthy

The new "hack" migrations will apply cleanly, and do not reference the
conflicting currencies table. No history will exist of the old
non-hack migrations, and so the final hack commit will not change
history.

5) Nothing has run, clean system

The new "hack" migrations will apply cleanly, and no history will be
re-written. The currencies table is never managed by haulage, and can
be cleanly created by CoLTE.

Important edge case
-------------------

This commit does *not* address the case that a subset of the orignal
haulage migrations were applied and then CoLTE was run. There was only
a very brief public release window where this could be the case, and
the CoLTE migrations would then fail to run due to the conflict on the
currencies table. At the point the currencies table is dropped by
later haulage migrations currently released, the system should end up
in state 1.

It is possible for users to end up in a bad state if they skip
straight from haualge 0.2.0 to this release (likely 0.2.7), where the
currencies table was created by haulage but has not yet been dropped
by haulage, but this is incredibly unlikely due to the timing of the
releases, and will be covered in the release notes.