haulage: Assume no migrations on applied migration query failure
If the database has not had any migrations run, there may not even be
an _sqlx_migrations table to query, which will result in an
error. With db-auto-update (i.e. auto migrate) it is possible that the
binary is being run for the first time on a fresh db, and all
migrations should be applied.