haulage: Add ratelimit skeleton
This commit adds a rough skeleton of where the ratelimit code will
eventually integrate into the haulage enforcer. The only ratelimit
policy supported right now is "unlimited", and the code currently
doesn't even ensure that no limit is set.

This commit also adds database migrations to add the ratelimit
information to each subscriber tuple in the backing database. The
current design makes a table of available policies, and then allocates
a jsonb column for the uplink and downlink parameters for each
subscriber. This allows the user-facing code to customize the limit
for each subscriber, although in practice it probably makes sense for
the user-facing code to allow applying a smaller set of named policies
across many users instad of configuring each user separately.