haulage: Add support for a token-bucket ratelimit policy
This commit adds an implementation of a simple token bucket-based
ratelimit policy. The token bucket qdisc is inserted before each
subscriber's stochastic fair queueing qdisc. Currently the only
exposed parameter is the rate, and the burst and latency are derived
from the rate with a target additional latency of <20ms. This is a
somewhat fast time constant depending on the nature of the backhaul
network, so we might want to eventually revisit it and allow it to be
configured.