haulage: Remove unecessary qdisc deletes
tc defines that removing a root qdisc will remove all child classes,
filters, and qdiscs. Instead of manually parsing them all and then
removing them, we can just remove the root's child to clear the
current qdisc configuration on the interface with only one tc
command. We still need to examine the configuration of the interface
to see if a non-default qdisc is present or not, since an error will
be generated if we try to remove the 0: qdisc.