haulage: Use more semantically meaningful conversions
I was orginally sloppy when implementing string conversions, and used
to_string in many places where it was not necessary or where it was
not meaningfully describing the conversion that was actually taking
place. This commit cleans this up, using as_str or to_owned in many
cases where we're not actually reformatting a std::String as a string.