haulage: Add upstream interface configuration
Adding the ability to configure both the upstream (and implicitly
assumed subscriber interface) will allow haulage to regulate both
uplink and downlink traffic rates. Previously there was only one
"interface" configured, referring to the interface with the subscriber
traffic. This should now be the more explicit "subscriberInterface" in
the config.

This commit includes logic to maintain backwards compatibility with
legacy configs with only a single "interface" configuration specified.