haulage: Add a catchall queue to upstream interface
Since QFQ will only forward packets that match, we need to add a
matchall class to allow traffic from the local host to egress the
upstream interface. Technically we should probably limit the allowed
egress to packets actually from the local host, but this can be
addressed more easily when using nf_tables for classification in the
future rather than trying to maintain the tc match rules.