haulage: Prefer deleting when changing user qdiscs
Only some qdiscs can be used with replace, and replacing can lead to
not resetting counters when restarting the system if qdiscs are not
changed. This commit changes the user modification commands to
explicilty delete and then add qdiscs when modifying the user's limit
configuration.

To support deleting the qdisc direclty attached to the user class, it
replaces the class qdisc (which is initialized with default handle 0:)
with a simple temporary pfifo qdisc. This is almost immediately delted
when the user's actual limit configuration is applied, but allows the
limit configurations to consistently delete before adding without
worrying about a special case at startup.

This commit establishes the convention that the leaf qdisc for each
user has a handle starting with "0xA", which makes it easier to see
them in the tc output and should give a consistent target for future
classification rules to direct packets.