haulage: Unify bridge enforcement with general access policies
This commit updates the enforcement logic which was previously
determining if the user was bridged or not bridged, and unifies it
with the new more general access policy capability. A user could have
a no balance access policy that specified their rate is limited, or
could have a policy that specifies they are blocked.