haulage: Track currently applied policy for update polling
Since the policy can currently be updated asynchronously in the
database by the colte UI, this commit adds logic to track the
currently applied policy so database queries can be made to determine
when the enforcement logic should be updated. Tracking the current
policy in the database allows queries to detect a mismatch between the
current policy and the expected policy given the business logic
condition of each subscriber directly in the database query. This
avoids needing to poll the configuration of each subscriber each poll
interval, only returning the information of the subscribers whose
state has changed.