haulage: Consistently return errors when setting policy
This commit makes the error handling consistent for all the sub-calls
in the top-level set policy function. All errors are returned early
and Ok is only returned if the entire set completes successfully. The
database state update is moved to the end of the function to allow
recovery in case the set fails. In that case the current database
state will not refelect the desired state, and the periodic polling
function should cause another update.

All update commands need to be idempotent, for this to work well, and
currently are implemented idempotently.