haulage: More robustly validate nonroot qdiscs
The original implementation assumed there could only be one item
returned by qdisc show when there was only a root qdisc, but some
classful qdiscs can also report class qdiscs as part of the default
qdisc. This commit changes the implementation to check if a qdisc can
be found with a non-root handle.