haulage: Re-work to use htb instead of QFQ+tbf
There is some kind of interaction between QFQ and the token bucket
implementation that I cannot figure out, but that causes kernel panics
in debian buster on x86_64 in deployment. For now we can avoid the
complexity of qfq by shifting to a purely htb-based
implementation. This implementation relies on the borrowing semantics
of htb, essentially setting up unlimited flows with a default 100kbps
rate, but then allowing them to have a very high ceiling. If a network
were be configured with "limited" flows with a rate much higher than
100kbps and the network were saturated, the limited flows would
actually get more throughput than the "unlimited" flows due to the
sharing rates being allocated proportionally according to the base
rate.

A future commit will address this issue by setting the base rate of
limited flows according to the min of their configured rate and the
base sharing rate shared by all "unlimited" flows.